refactor(C5): extract _RouterUsageShim → runtime/usage_shim.py (#1792)
FP-0044 C-series stage 5 — the final non-Session class extraction, completing
Stage-1 (session.py god-file class decomposition). Extracts the
`_RouterUsageShim` token-usage adapter verbatim from `reyn.runtime.session`
into a new `reyn.runtime.usage_shim` module.

- Pure move, byte-identical class body (verified removed==added via diff).
- `usage_shim.py` docstring is current-state only; story in this commit
  (C-series discipline, C1 #1790 R1).
- Dependency-free wrapper (`usage: object`); no imports needed, no cycle.
- session.py does NOT reference the shim internally (only router_loop_driver
  does, function-locally), so no back-import is added — session.py simply
  drops the class.
- Clean break, no shim: both consumers repointed
  `from reyn.runtime.session import _RouterUsageShim`
  → `from reyn.runtime.usage_shim import _RouterUsageShim` (mixed lines split
  so `Session` / `_render_summary_for_storage` stay sourced from session).
- Repo-wide straggler 0 across all ref classes incl multi-line parenthesised
  import blocks (C4 learning applied up-front).
